Renames Region.Policy into Region.Props
Also splits Region.scala into smaller files.
